requirements.txt
altair==5.5.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
confection==0.1.5
cryptography==45.0.3
cymem==2.0.11
Django==5.2.2
django-widget-tweaks==1.5.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langcodes==3.5.0
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
narwhals==1.41.0
numpy==2.2.6
packaging==24.2
pandas==2.2.3
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
preshed==3.0.10
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
streamlit==1.45.1
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
whitenoise==6.9.0
wrapt==1.17.2
